+++ CHANGES.md
@@ -7,6 +7,7 @@
 
 ### Fixes
 - Do not count documentation length when aligning all columns in settings section  [#156](https://github.com/MarketSquare/robotframework-tidy/issues/156)
+- Acknowledge ``--lineseparator`` option [#163](https://github.com/MarketSquare/robotframework-tidy/issues/163)
 
 ## 1.4.0